Fixes various bugs with log and search
* Fixes incorrect iteration offset use in text_indexer.c
* Makes index lookups non blocking and async
* Clarifies logging usage
* Fixes memory leak in newline tokenization

+++ src/index.h
@@ -52,7 +52,7 @@ int f_index_init(f_index** out, char* filename, int filename_len, f_lookup_file*
   @param size an int ref that is populated with the size of the fetched string
   @return non zero for error
 */
-int f_index_lookup(char** out, f_index* index, unsigned int start, unsigned int count, int* size);
+int f_index_lookup(char** out, f_index* index, size_t start, size_t count, int* size);
 
 /**
   Frees an index and it's lookup
